refactor(slack): simplify fenced section slicing
