+++ packages/shared/src/slack/sections.ts
@@ -166,26 +166,29 @@ class SlackSectionAccumulator {
 
   private takeHardSlice(text: string): string {
     const start = this.sectionEnd;
-    let budget = this.maxChars - reopenPrefix(start).length;
-    while (budget > 0) {
-      const taken = sliceAtCodePointBoundary(text, budget);
-      if (!taken) break;
-
-      const end = advanceFence(start, taken);
-      const renderedLength = reopenPrefix(start).length + taken.length + closeSuffix(end).length;
-      if (renderedLength <= this.maxChars) {
-        this.sectionStart = start;
-        this.body = taken;
-        this.sectionEnd = end;
-        return taken;
-      }
-
-      // Fence state must be measured at the actual cut: a token can close a
-      // fence after this slice, while this section still needs a closing repair.
-      budget -= renderedLength - this.maxChars;
+    const prefix = reopenPrefix(start);
+    let taken = sliceAtCodePointBoundary(text, Math.max(0, this.maxChars - prefix.length));
+    if (!taken) {
+      throw new RangeError("Section budget is too small to fit the next Unicode code point");
     }
 
-    throw new RangeError("Section budget is too small to fit the next Unicode code point");
+    let end = advanceFence(start, taken);
+    const suffix = closeSuffix(end);
+    if (prefix.length + taken.length + suffix.length > this.maxChars) {
+      taken = sliceAtCodePointBoundary(
+        text,
+        Math.max(0, this.maxChars - prefix.length - suffix.length)
+      );
+      if (!taken) {
+        throw new RangeError("Section budget is too small to fit the next Unicode code point");
+      }
+      end = advanceFence(start, taken);
+    }
+
+    this.sectionStart = start;
+    this.body = taken;
+    this.sectionEnd = end;
+    return taken;
   }
 
   private flush(): void {
