+++ packages/slack-bot/src/completion/blocks.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from "vitest";
 import { buildCompletionBlocks } from "./blocks";
-import { splitIntoSlackSections } from "@open-inspect/shared/slack";
 import type { AgentResponse, SlackCallbackContext } from "../types";
 
 const BASE_CONTEXT: SlackCallbackContext = {
@@ -226,44 +225,6 @@ describe("long response handling", () => {
     }
   });
 
-  it("preserves whitespace exactly across section boundaries", () => {
-    const textContent = `\n  alpha\n\n\n\n\nbeta\n\n${"x".repeat(4000)}  \n`;
-    const sections = splitIntoSlackSections(textContent);
-
-    expect(sections.join("")).toBe(textContent);
-  });
-
-  it("keeps Unicode code points intact across hard section boundaries", () => {
-    const textContent = `${"a".repeat(2999)}😀${"b".repeat(10)}`;
-    const sections = splitIntoSlackSections(textContent);
-
-    expect(sections.join("")).toBe(textContent);
-    for (const section of sections) {
-      const firstCodeUnit = section.charCodeAt(0);
-      const lastCodeUnit = section.charCodeAt(section.length - 1);
-      expect(firstCodeUnit >= 0xdc00 && firstCodeUnit <= 0xdfff).toBe(false);
-      expect(lastCodeUnit >= 0xd800 && lastCodeUnit <= 0xdbff).toBe(false);
-    }
-  });
-
-  it("rejects a section budget that cannot fit the next Unicode code point", () => {
-    expect(() => splitIntoSlackSections("😀", 1)).toThrow(
-      new RangeError("Section budget is too small to fit the next Unicode code point")
-    );
-  });
-
-  it("keeps Unicode code points intact when adding the truncation marker", () => {
-    for (let prefixLength = 2900; prefixLength <= 3000; prefixLength += 1) {
-      const textContent = `${"a".repeat(prefixLength)}😀${"b".repeat(4000)}\n\nmore`;
-      const [section] = splitIntoSlackSections(textContent, 3000, 1);
-      const markerIndex = section.indexOf("_...truncated");
-      expect(markerIndex).toBeGreaterThanOrEqual(0);
-      const content = section.slice(0, markerIndex).trimEnd();
-      const lastCodeUnit = content.charCodeAt(content.length - 1);
-      expect(lastCodeUnit >= 0xd800 && lastCodeUnit <= 0xdbff).toBe(false);
-    }
-  });
-
   it("never exceeds Slack's per-section character limit", () => {
     const textContent = "x".repeat(25_000);
     const blocks = buildCompletionBlocks(
@@ -307,19 +268,6 @@ describe("long response handling", () => {
     }
   });
 
-  it("balances fences that open and close on the same line", () => {
-    const fence = "```";
-    const sections = splitIntoSlackSections(`${fence}code${fence}\n${"after ".repeat(700)}`);
-
-    expect(sections.length).toBeGreaterThan(1);
-    for (const section of sections) {
-      expect((section.match(/```/g) ?? []).length % 2).toBe(0);
-    }
-  });
-
-  // The two checks above pass on fence-free and short-line input respectively, so
-  // neither exercises a split *inside* a fence — which is where the section repair
-  // adds characters after the fit check and where the hard slice drops them.
   it("respects the section cap when a fence is split (repair chars are budgeted)", () => {
     const textContent = `\`\`\`json\n${"a".repeat(9000)}\n\`\`\``;
     const blocks = buildCompletionBlocks(
@@ -330,73 +278,6 @@ describe("long response handling", () => {
     );
     for (const text of sectionTexts(blocks)) {
       expect(text.length).toBeLessThanOrEqual(3000);
-    }
-  });
-
-  it("respects the section cap when one oversized token opens and closes a fence", () => {
-    const sections = splitIntoSlackSections(`\`\`\`${"x".repeat(4000)}\`\`\``);
-
-    expect(sections.length).toBeGreaterThan(1);
-    for (const section of sections) {
-      expect(section.length).toBeLessThanOrEqual(3000);
-      expect((section.match(/```/g) ?? []).length % 2).toBe(0);
-    }
-  });
-
-  it("loses no characters when hard-slicing inside a fence", () => {
-    const payload = "b".repeat(7000);
-    const sections = splitIntoSlackSections(`\`\`\`js\n${payload}\n\`\`\``);
-    const recovered = sections
-      .join("")
-      .split("```")
-      .join("")
-      .replace(/^js$/gm, "")
-      .replace(/\n/g, "");
-    expect(recovered).toBe(payload);
-  });
-
-  it("keeps fences balanced in the truncated final section", () => {
-    const textContent = `\`\`\`ts\n${Array.from({ length: 400 }, () => "y".repeat(2900)).join("\n")}\n\`\`\``;
-    const sections = splitIntoSlackSections(textContent);
-    const last = sections[sections.length - 1];
-    expect(last).toContain("truncated");
-    expect(last.length).toBeLessThanOrEqual(3000);
-    for (const section of sections) {
-      expect((section.match(/```/g) ?? []).length % 2).toBe(0);
-    }
-  });
-
-  // The truncation cut can land inside a fence the final section both opened and
-  // closed, which a trailing-fence check cannot detect. Sweep the section length
-  // through the window where slicing actually happens, moving the fence across the
-  // cut point, and assert the invariant holds for every shape.
-  it("keeps fences balanced when the truncation cut lands inside a fence", () => {
-    const fence = "```";
-    for (let sectionLen = 2949; sectionLen <= 3000; sectionLen += 1) {
-      for (const codeLen of [20, 45, 200]) {
-        const block = `\n${fence}ts\n${"h".repeat(codeLen)}\n${fence}`;
-        const fill = sectionLen - block.length;
-        if (fill < 1) continue;
-        const paragraphs = [
-          ...Array.from({ length: 19 }, () => "f".repeat(2900)),
-          "g".repeat(fill) + block,
-          ...Array.from({ length: 5 }, () => "z".repeat(2900)),
-        ];
-        const sections = splitIntoSlackSections(paragraphs.join("\n\n"));
-        const last = sections[sections.length - 1];
-        expect(last).toContain("truncated");
-        expect(last.length).toBeLessThanOrEqual(3000);
-        expect((last.match(/```/g) ?? []).length % 2).toBe(0);
-      }
-    }
-  });
-
-  it("carries the fence language across a split", () => {
-    const sections = splitIntoSlackSections(`\`\`\`python\n${"c".repeat(6500)}\n\`\`\``);
-    expect(sections.length).toBeGreaterThan(1);
-    // Every continuation section reopens the fence with its original language.
-    for (const section of sections.slice(1)) {
-      expect(section.startsWith("```python\n")).toBe(true);
     }
   });
 
@@ -433,44 +314,5 @@ describe("long response handling", () => {
       "https://inspect.example.com"
     );
     expect(sectionTexts(blocks)[0]).toBe("_Agent completed._");
-  });
-});
-
-describe("fence info bounding", () => {
-  // Regression: `info` was captured unbounded from the fence line, so an
-  // oversized fence-opener (a single ≥3000-char line beginning with ```) was
-  // re-emitted by reopenPrefix on every continuation section and overflowed
-  // Slack's per-section cap. Slack rejects the whole message on overflow rather
-  // than trimming the block, so the cap has to hold for every input shape.
-  it("keeps sections within the cap for an oversized fence-opener line", () => {
-    const monsterInfo = "x".repeat(4000);
-    const sections = splitIntoSlackSections(`\`\`\`${monsterInfo}\n${"c".repeat(5000)}\n\`\`\``);
-    for (const section of sections) {
-      expect(section.length).toBeLessThanOrEqual(3000);
-    }
-  });
-
-  it("keeps only the language token when the fence line carries trailing text", () => {
-    const sections = splitIntoSlackSections(
-      `\`\`\`python title="a very long annotation ${"y".repeat(200)}"\n${"c".repeat(6500)}\n\`\`\``
-    );
-    expect(sections.length).toBeGreaterThan(1);
-    for (const section of sections.slice(1)) {
-      expect(section.startsWith("```python\n")).toBe(true);
-    }
-  });
-
-  // The bot's review caught the original overflow by fuzzing rather than by a
-  // single case, and a single case would have missed it here too.
-  it("never overflows across a sweep of fence-opener and body lengths", () => {
-    for (let infoLen = 0; infoLen <= 4200; infoLen += 350) {
-      for (let bodyLen = 2900; bodyLen <= 6200; bodyLen += 550) {
-        const text = `\`\`\`${"i".repeat(infoLen)}\n${"b".repeat(bodyLen)}\n\`\`\``;
-        for (const section of splitIntoSlackSections(text)) {
-          expect(section.length).toBeLessThanOrEqual(3000);
-          expect((section.match(/```/g) ?? []).length % 2).toBe(0);
-        }
-      }
-    }
   });
 });
